render option added to gui

+++ source/scene/camera.h
@@ -43,11 +43,11 @@ public:
 	float MouseSensitivity;
 	float Zoom;
 
-	unsigned int screenWidth;
-	unsigned int screenHeight;
+	size_t screenWidth;
+	size_t screenHeight;
 
 	// Constructor with vectors
-	Camera(unsigned int _screenWidth, unsigned int _screenHeight, 
+	Camera(size_t _screenWidth, size_t _screenHeight, 
 	glm::vec3 position = glm::vec3(0.0f, 0.0f, 3.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH) : 
 	Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY), Zoom(ZOOM)
 	{
@@ -60,7 +60,7 @@ public:
 		updateCameraVectors();
 	}
 	// Constructor with scalar values
-	Camera(unsigned int _screenWidth, unsigned int _screenHeight, 
+	Camera(size_t _screenWidth, size_t _screenHeight, 
 	float posX, float posY, float posZ, float upX, float upY, float upZ, float yaw, float pitch) : 
 	Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY), Zoom(ZOOM)
 	{
